Refactored Client/ For Future Work Involving Redux & Unit Testing
