+++ client/src/components/App.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <h1>Hello From TeleQuest</h1>
+      </div>
+    );
+  }
+}
+
+export default App;
